- Removed redundant CoreModule. - Moved services from "core" to each corresponding "feature" folder.
