+++ src/app/home/movies/services/movie.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { environment } from '../../../environments/environment';
-import { Movie } from '../models/movie.model';
+import { environment } from '../../../../environments/environment';
+import { Movie } from '../../../core/models/movie.model';
 
 @Injectable({
   providedIn: 'root'
