+++ src/app/home/schedules/schedule/schedule.component.ts
@@ -1,10 +1,10 @@
 import { Component, OnInit, Output, EventEmitter, Inject } from '@angular/core';
 import { MAT_DIALOG_DATA } from '@angular/material';
 import { ScheduleRequest, Schedule } from '../../../core/models/schedule.model';
-import { TheaterService } from '../../../core/services/theater.service';
+import { TheaterService } from '../../theaters/services/theater.service';
 import { Movie } from '../../../core/models/movie.model';
-import { MovieService } from '../../../core/services/movie.service';
-import { ScheduleService } from '../../../core/services/schedule.service';
+import { MovieService } from '../../movies/services/movie.service';
+import { ScheduleService } from '../services/schedule.service';
 import { Theater } from '../../theaters/models/theater.model';
 
 @Component({
